View: move component specific logic from AbstractView to ComponentView
