+++ src/Application.ts
@@ -36,7 +36,7 @@ export class Application
 	public run(controller: any): void
 	{
 		setTimeout(() => {
-			let view = new ApplicationView(<any>document, controller);
+			let view = new ApplicationView(this.container, <any>document, controller);
 			this.compiler.compile(view);
 			view.watcher.run();
 		}, 0);
