+++ tests/mocks/MockApplicationView.ts
@@ -1,0 +1,14 @@
+import {ApplicationView} from '../../core';
+import {Container} from '../../di';
+
+
+export class MockApplicationView extends ApplicationView
+{
+
+
+	constructor(container: Container)
+	{
+		super(container, <any>{}, <any>{});
+	}
+
+}
