--- tests/mocks/MockView.ts
@@ -1,13 +1,0 @@
-import {AbstractView} from '../../src/Views/AbstractView';
-
-
-export class MockView extends AbstractView
-{
-
-
-	constructor()
-	{
-		super();
-	}
-
-}
