+++ tests/tests/Templating/Binding/PropertyBinding.ts
@@ -3,13 +3,14 @@ import {Container} from '../../../../di';
 import {Dom} from '../../../../utils';
 import {PropertyBinding} from '../../../../src/Templating/Binding/PropertyBinding';
 import {ExpressionParser} from '../../../../src/Parsers/ExpressionParser';
-import {MockView} from '../../../mocks/MockView';
+import {MockApplicationView} from '../../../mocks/MockApplicationView';
 
 import chai = require('chai');
 
 
 let expect = chai.expect;
 
+let container: Container = null;
 let application: Application = null;
 let compiler: Compiler = null;
 
@@ -17,7 +18,7 @@ let compiler: Compiler = null;
 describe('#Templating/Binding/PropertyBinding', () => {
 
 	beforeEach(() => {
-		let container = new Container;
+		container = new Container;
 		application = new Application(container);
 		compiler = container.get(<any>Compiler);
 	});
@@ -29,7 +30,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hello: 'good day',
 			});
 
@@ -44,7 +45,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				obj: {
 					greetings: ['good day'],
 				},
@@ -61,7 +62,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				a: 1,
 				b: 2,
 				c: 3,
@@ -76,7 +77,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 		it("should bind value to element's inner html", () => {
 			let el = Dom.el('<div [innerHTML]="hello"></div>');
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hello: 'good day',
 			});
 
@@ -91,7 +92,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el.style.border).to.be.equal('');
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				border: '1px solid red',
 			});
 
@@ -106,7 +107,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el.style.border).to.be.equal('1px solid red');
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				border: false,
 			});
 
@@ -121,7 +122,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el.classList.contains('alert')).to.be.equal(false);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hasAlert: true,
 			});
 
@@ -136,7 +137,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el.classList.contains('alert')).to.be.equal(true);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hasAlert: false,
 			});
 
@@ -151,7 +152,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hello: 'good day',
 			});
 
@@ -175,7 +176,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				a: [
 					{
 						b: 'good day',
@@ -207,7 +208,7 @@ describe('#Templating/Binding/PropertyBinding', () => {
 
 			expect(el['test']).to.be.equal(undefined);
 
-			let view = new ComponentView(new MockView, new ElementRef(el), {
+			let view = new ComponentView(new MockApplicationView(container), new ElementRef(el), {
 				hello: 'good day',
 			});
